added

+++ northwind/northwind/src/main/java/kodlamaio/northwind/entities/concretes/Category.java
@@ -28,7 +28,7 @@ public class Category {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="category_id")
-	private short categoryId;
+	private int categoryId;
 	
 	
 	@Column(name="category_name")
